pyproject.toml
[tool.poetry]
name = "dagster_pipeline"
version = "0.1.0"
description = ""
authors = ["Heiko Kulinna <Heikokulinna@gmx.de>"]

packages = [{ include = "pipelines" }]


[tool.poetry.dependencies]
python = ">=3.8, <3.9"
pandas = "^1.3.4"
numpy = "^1.21.4"
azure-storage-blob = "12.7.1"
cryptography = "<= 3.4.7"

# make it work again
protobuf = "< 3.20.0"
jinja2 = "< 2.11"
MarkupSafe = "< 2.1"
itsdangerous = "< 2.1"
ipython_genutils = "*"   # 0.2.0


dagster = "0.13.5"
dagster-docker = "0.13.5"
dagster-postgres = "0.13.5"
dagster-k8s = "0.13.5"
dagit = "0.13.5"
python-dotenv = "^0.19.2"
"yaml-1.3" = ">= 0.1.0"

pyarrow = "^6.0.1"
dask = {extras = ["distributed"], version = "^2022.1.0"}
dask-azureblobfs = "^0.1.0"
adlfs = "^2021.10.0"



[tool.poetry.dev-dependencies]
black = "^21.12b0"
flake8 = "^4.0.1"
isort = "^5.10.1"
pre-commit = "^2.16.0"
pylint = "^2.12.2"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"




